requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pydantic==2.5.0

# Data & Analysis
pandas==2.1.4
numpy==1.25.2
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2

# AI & ML
groq==0.4.1
openai==1.3.7
langchain==0.1.0
transformers==4.36.2
torch==2.1.2

# Database & Caching
sqlalchemy==2.0.23
redis==5.0.1
psycopg2-binary==2.9.9

# Visualization & Charts
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Financial Analysis
ta==0.10.2
scipy==1.11.4
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
python-multipart==0.0.6

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0
